airflow/requirements.txt
dbt-bigquery==1.11.1
apache-airflow-providers-google==10.13.1